cfanatics.html edits - intro text
added and formatted the intro, overview, and overview grid items in the cfanatics page.
added a new item in the project_page.css for a grid that is 2 columns on mobile, but 3 on desktop.
also eddited the project page css a bit more to make sure that the text wdith was 750px as intended.